Update imgsizefix.coffee
Sometimes you want to load a image after everything else is loaded. In that case you have to store the real image url in a different parameter than "src". This change offers a lazy property. Its not a good solution that can be used in any situation but it does solve my current problem.